+++ src/app/components/investment-result/investment-result.component.ts
@@ -1,4 +1,4 @@
-import { formatNumber } from '@angular/common';
+import { formatNumber, formatDate } from '@angular/common';
 import { Component, Inject, Input, LOCALE_ID, OnChanges, SimpleChanges } from '@angular/core';
 import { GlobalMethodsService } from 'src/app/services/global-methods.service';
 import { InvestingDatum } from 'src/shared-class/InvestingDatum';
@@ -14,9 +14,7 @@ export class InvestmentResultComponent implements OnChanges {
   @Input() data!: InfoForStatistic[];
 
   statistics!: ViewDatum[];
-
   displayedColumnInfos!: ViewColumnInfo[];
-
   get displayedColumns(): String[] {
     return this.displayedColumnInfos.map(i => i.strIndex);
   }
@@ -30,42 +28,65 @@ export class InvestmentResultComponent implements OnChanges {
   constructor(@Inject(LOCALE_ID) private locale: string, private gvs: GlobalMethodsService) { }
 
   ngOnChanges(changes: SimpleChanges) {
-    this.statistics = this.getStatistics(changes.data.currentValue);
+
+    this.statistics = [];
+    this.displayedColumnInfos = [];
+    this.displayedData = [];
+
+    if (changes.data !== undefined && changes.data.currentValue !== undefined) {
+
+      let results = this.getStatistics(changes.data.currentValue);
+
+      this.statistics = results.statistics;
+      this.displayedColumnInfos = results.displayedColumnInfos;
+      this.displayedData = results.displayedData;
+    }
   }
 
-  getStatistics(data: InfoForStatistic[]): ViewDatum[] {
+  getStatistics(data: InfoForStatistic[]): {
+    statistics: ViewDatum[],
+    displayedColumnInfos: ViewColumnInfo[],
+    displayedData: any[],
+  } {
 
     let _this = this;
 
-    let viewData: ViewDatum[] = [ViewDatum.createHeader()];
+    let statistics: ViewDatum[] = [ViewDatum.createHeader()];
     for (let datum of data) {
 
       let vd = ViewDatum.createByInfo(
         datum,
-        (value: number, digitsInfo?: string): string | undefined => {
+        (value: number, digitsInfo?: string): string => {
           return formatNumber(value, _this.locale, digitsInfo);
+        },
+        (value: Date): string => {
+          return formatDate(value, 'yyyy/MM/dd', _this.locale);
         });
 
       //console.log(vd);
 
-      viewData.push(vd);
+      statistics.push(vd);
     }
 
-    this.displayedColumnInfos = viewData.map(datum => new ViewColumnInfo(datum.title, datum.index));
-    this.displayedData =
+    let displayedColumnInfos = statistics.map(datum => new ViewColumnInfo(datum.title, datum.index));
+    let displayedData =
       Object.getOwnPropertyNames(new ViewDatum())
         .filter(p => p !== 'title' && p !== 'index') // already used
         .map(prop => {
           const output: any = {};
 
-          for (let i = 0; i < viewData.length; i++) {
-            output[viewData[i].index] = viewData[i][prop as keyof ViewDatum];
+          for (let i = 0; i < statistics.length; i++) {
+            output[statistics[i].index] = statistics[i][prop as keyof ViewDatum];
           }
 
           return output;
         });
 
-    return viewData;
+    return {
+      statistics,
+      displayedColumnInfos,
+      displayedData,
+    };
   }
 }
 
@@ -86,13 +107,13 @@ class ViewColumnInfo {
 }
 
 class ViewDatum {
-  title: string = '';
-  description: string = '';
-  count: string = '';
-  maxMinCost: string = '';
-  totalCost: string = '';
-  currentValue: string = '';
-  gain: string = '';
+  title: string = '--';
+  description: string = '--';
+  dateRange: string = '--';
+  maxMinCost: string = '--';
+  totalCost: string = '--';
+  currentValue: string = '--';
+  gain: string = '--';
 
   index: number = -1;
   static global_index: number = 0;
@@ -103,8 +124,19 @@ class ViewDatum {
     this.index = ViewDatum.global_index++;
   }
 
-  static createByInfo(datum: InfoForStatistic, formatNumber: any) {
+  static createByInfo(datum: InfoForStatistic, formatNumber: any, formatDate: any) {
+
+    let rtn = new ViewDatum({
+      title: datum.title,
+      description: 'Invest on the 1st of each month',
+    });
+
     let info = datum.info;
+    let count = info.length;
+
+    if (count === 0) {
+      return rtn;
+    }
 
     let maxCostPerUnit = Number.MIN_VALUE;
     let minCostPerUnit = Number.MAX_VALUE;
@@ -118,7 +150,18 @@ class ViewDatum {
       }
     });
 
-    let count = info.length;
+    let maxDate = new Date(-8640000000000000);
+    let minDate = new Date(8640000000000000);
+    info.forEach(i => {
+      let d = i.tradeDate;
+      if (d > maxDate) {
+        maxDate = d;
+      }
+      if (d < minDate) {
+        minDate = d;
+      }
+    });
+
     let lastInfo = info[count - 1];
     let currentValue = lastInfo.currentValue;
     let totalCost = lastInfo.cumulativeDollarCost;
@@ -128,7 +171,7 @@ class ViewDatum {
     return new ViewDatum({
       title: datum.title,
       description: 'Invest on the 1st of each month',
-      count: formatNumber(count),
+      dateRange: formatDate(minDate) + '-' + formatDate(maxDate) + ' (' + formatNumber(count) + ' months)',
       maxMinCost: formatNumber(maxCostPerUnit, '1.0-0') + '/' + formatNumber(minCostPerUnit, '1.0-0'),
       currentValue: formatNumber(currentValue, '1.0-0'),
       totalCost: formatNumber(totalCost, '1.0-0'),
@@ -140,7 +183,7 @@ class ViewDatum {
     return new ViewDatum({
       title: '',
       description: 'Description',
-      count: 'Number of months',
+      dateRange: 'Date Range',
       maxMinCost: 'Max/Min Investment of months',
       currentValue: 'Current Value',
       totalCost: 'Total Investment',